+++ tests/Integration/QueryValidationTest.php
@@ -156,6 +156,13 @@ class QueryValidationTest extends TestCase
     public function invalidProvider(): array
     {
         return [
+            'fields:not array' => [
+                'fields',
+                'blah',
+                'The fields must be an array.',
+                'fields',
+                ['rule' => 'array'],
+            ],
             'fields:singular' => [
                 'fields',
                 ['posts' => 'title,content'],
@@ -184,6 +191,13 @@ class QueryValidationTest extends TestCase
                 'fields',
                 ['rule' => 'allowed-field-sets'],
             ],
+            'filter:not array' => [
+                'filter',
+                'blah',
+                'The filter must be an array.',
+                'filter',
+                ['rule' => 'array'],
+            ],
             'filter:singular' => [
                 'filter',
                 ['foo' => 'bar'],
@@ -197,6 +211,13 @@ class QueryValidationTest extends TestCase
                 'Filter parameters baz, foo are not allowed.',
                 'filter',
                 ['rule' => 'allowed-filter-parameters'],
+            ],
+            'include:not string' => [
+                'include',
+                ['foo' => 'bar'],
+                'The include must be a string.',
+                'include',
+                ['rule' => 'string'],
             ],
             'include:singular' => [
                 'include',
@@ -212,6 +233,13 @@ class QueryValidationTest extends TestCase
                 'include',
                 ['rule' => 'allowed-include-paths'],
             ],
+            'page:not array' => [
+                'page',
+                'blah',
+                'The page must be an array.',
+                'page',
+                ['rule' => 'array'],
+            ],
             'page:singular' => [
                 'page',
                 ['number' => '1', 'size' => '25', 'foo' => 'bar'],
@@ -226,6 +254,13 @@ class QueryValidationTest extends TestCase
                 'page',
                 ['rule' => 'allowed-page-parameters'],
             ],
+            'sort:not string' => [
+                'sort',
+                ['foo' => 'bar'],
+                'The sort must be a string.',
+                'sort',
+                ['rule' => 'string'],
+            ],
             'sort:singular' => [
                 'sort',
                 'title,createdAt,foo',
@@ -239,6 +274,13 @@ class QueryValidationTest extends TestCase
                 'Sort parameters bar, foo are not allowed.',
                 'sort',
                 ['rule' => 'allowed-sort-parameters'],
+            ],
+            'withCount:not string' => [
+                'withCount',
+                ['foo' => 'bar'],
+                'The with count must be a string.',
+                'withCount',
+                ['rule' => 'string'],
             ],
             'withCount:singular' => [
                 'withCount',